fix: harden bf-tree no_std for bare-metal targets
- Make io-uring, libc, windows-sys optional deps gated behind std feature
  via target-conditional [dependencies] sections
- Gate OS-specific atomic_wait platforms behind feature = "std", falling
  back to spin-loop on bare-metal or no_std builds
- Add portable-atomic for AtomicU64 on 32-bit targets (thumbv7em)
- Fix spdk_vfs cfg gates to require target_os = "linux"
- Add MappingTable debug_assert for batch_id bounds + Vec::with_capacity
- Document single-core spin deadlock risk and 4096-byte alignment requirement
- Add no_std smoke test (crates/bf-tree/tests/no_std_smoke.rs)
- Verified: cargo check --no-default-features --target thumbv7em-none-eabihf

+++ crates/bf-tree/src/utils/mod.rs
@@ -182,9 +182,9 @@ pub(crate) fn get_rng() -> SmallThreadRng {
 #[cfg(not(all(feature = "shuttle", test)))]
 #[cfg(not(feature = "std"))]
 pub(crate) fn get_rng() -> rand::rngs::SmallRng {
-    use core::sync::atomic::{AtomicU64, Ordering};
-    static SEED_CTR: AtomicU64 = AtomicU64::new(1);
-    rand::rngs::SmallRng::seed_from_u64(SEED_CTR.fetch_add(1, Ordering::Relaxed))
+    use core::sync::atomic::{AtomicU32, Ordering};
+    static SEED_CTR: AtomicU32 = AtomicU32::new(1);
+    rand::rngs::SmallRng::seed_from_u64(SEED_CTR.fetch_add(1, Ordering::Relaxed) as u64)
 }
 
 #[cfg(all(feature = "shuttle", test))]
